chore: add allowed extension to `png, jpg, jpeg`

+++ main.py
@@ -6,6 +6,12 @@ from keras.models import load_model
 from keras.preprocessing.image import img_to_array
 
 app = Flask(__name__)
+app.config['ALLOWED_EXTENSIONS'] = set(['png', 'jpg', 'jpeg'])
+
+
+def allowed_file(filename):
+    return '.' in filename and filename.rsplit('.', 1)[1] in app.config['ALLOWED_EXTENSIONS']
+
 
 # Load the pre-trained semaphore model
 model_path = 'model_ml.h5'
@@ -25,50 +31,66 @@ def preprocess_image(image_path):
     return img
 
 
-# Route for semaphore image classification
 @app.route('/classify_semaphore', methods=['POST'])
 def classify_semaphore():
     if 'image' not in request.files:
-        return jsonify({'error': 'No image uploaded'}), 400
+        response = {
+            'success': 'false',
+            'message': 'No image uploaded'
+        }
+        return jsonify(response), 400
 
     image_file = request.files['image']
 
-    # Save the uploaded image temporarily
-    temp_image_path = 'temp_image.jpg'
-    image_file.save(temp_image_path)
+    if image_file and allowed_file(image_file.filename):
+        # Save the uploaded image temporarily
+        temp_image_path = 'temp_image.jpg'
+        image_file.save(temp_image_path)
 
-    # Preprocess the image
-    preprocessed_image = preprocess_image(temp_image_path)
+        # Preprocess the image
+        preprocessed_image = preprocess_image(temp_image_path)
 
-    # Perform prediction
-    predictions = model.predict(preprocessed_image)
+        # Perform prediction
+        predictions = model.predict(preprocessed_image)
 
-    # Get the class with the highest probability
-    predicted_class_index = np.argmax(predictions[0])
-    predicted_class = classes[predicted_class_index]
+        # Get the class with the highest probability
+        predicted_class_index = np.argmax(predictions[0])
+        predicted_class = classes[predicted_class_index]
 
-    # Get the confidence of the prediction
-    confidence = float(predictions[0][predicted_class_index])
+        # Get the confidence of the prediction
+        confidence = float(predictions[0][predicted_class_index])
 
-    # Set a threshold for confidence (adjust as needed)
-    confidence_threshold = 0.5  # Example threshold value
+        # Set a threshold for confidence (adjust as needed)
+        confidence_threshold = 0.5  # Example threshold value
 
-    # Prepare JSON response
-    if confidence >= confidence_threshold:
-        response = {
-            'predicted_class': predicted_class,
-            'confidence': confidence
-        }
+        # Prepare JSON response
+        if confidence >= confidence_threshold:
+            response = {
+                'success': "true",
+                'message': {
+                    'predicted_class': predicted_class,
+                    'confidence': confidence
+                }
+            }
+        else:
+            response = {
+                'success': "false",
+                'message': {
+                    'predicted_class': 'Not a Semaphore',
+                    'confidence': confidence
+                }
+            }
+
+        # Remove the temporary image file
+        os.remove(temp_image_path)
+
+        return jsonify(response)
     else:
         response = {
-            'predicted_class': 'Not a Semaphore',
-            'confidence': confidence
+            "success": "false",
+            "message": "Invalid file format. Please upload a JPG, JPEG, or PNG image."
         }
-
-    # Remove the temporary image file
-    os.remove(temp_image_path)
-
-    return jsonify(response)
+        return jsonify(response), 400
 
 
 if __name__ == '__main__':
